Create Tableau_Tornadoes.md
Topic: Tornadoes in United States

In absolute numbers, the United States has the most tornadoes by far (more than 1,000 per year have been reported every year since 1990). It also has the most violent tornadoes (about 10 to 20 per year). Tornadoes of this intensity are very infrequent outside of the United States.
The objective of this project will be to show the impact of tornadoes that have occurred in the United States from 1950 to 2018, along with showcasing the ones that have been the deadliest. With the help of the visualization techniques learnt in the Visual Analytics course, my aim was to depict the number of fatalities caused and the economic damage caused by these tornadoes. Also, this project helps in an understanding of where tornadoes happen, indicators of damage, detecting patterns and weather conditions associated with tornadoes.