Netflix_ratings_prediction
Netflix is a streaming platform for various types of content in different languages. The goal of our team was to help Netflix identify what kind of content they should produce that will result in higher ratings, which is our target variable. If Netflix has content with high ratings that users want to watch, they need a model to help identify what content to stream. We as a team would try to find insights from the scraped Netflix data and help production studios with actionable recommendations to predict the score for the movie/TV show success on a scale of 10 (0-low, 10-high).

We also tried to find insights on the content type, based on which we could make recommendations to the movie/TV show makers so that they could shift their focus to licensed/produced content accordingly. Additionally, we wanted to find a correlation between ratings and other variables such as genre, rank, cast, episodes, runtime, and show type, the different columns in our dataset.

It is common for people to read the summary or plot before watching a show. We also performed text analysis on title, synopsis, and description related to the show's rating. Netflix's business model is to produce content that is exclusive to their website. To further understand if "Netflix originals" are more likely to have higher ratings, we merged another dataset that classifies the shows.